requirements.txt
async-generator==1.10
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
black==23.3.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
contourpy==1.0.6
cryptography==41.0.3
cycler==0.11.0
exceptiongroup==1.1.3
Flask==2.2.2
fonttools==4.38.0
h11==0.14.0
http-ece==1.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.6.2
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.23.5
outcome==1.2.0
packaging==23.1
paho-mqtt==1.6.1
pathspec==0.11.2
Pillow==9.3.0
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
py-vapid==1.9.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.16.1
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pywebpush==1.14.0
pyzmq==24.0.1
redis==5.0.0
requests==2.31.0
rich==13.6.0
rq==1.15.1
ruff==0.0.277
scikit-learn==1.2.0
scipy==1.9.3
shellingham==1.5.3
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
statemachine==0.1
threadpoolctl==3.1.0
tk==0.1.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.9.2
typing_extensions==4.8.0
urllib3==2.0.4
uWSGI==2.0.21
wcwidth==0.2.8
websockets==11.0.3
websockets-cli==0.2.1
Werkzeug==2.2.2
wsproto==1.2.0
zmcat==0.0.11
